cli: Add socket-lb field
This commit adds the socket-lb field (aka host-reachable svc) to the
verbose cilium cli status output. E.g.:

    $ cilium status --verbose
    [..]
    KubeProxyReplacement Details:
      Status:              Strict
      Protocols:           TCP, UDP
      Devices:             wlp3s0 192.168.178.29 (Direct Routing)
      Mode:                SNAT
      Backend Selection:   Random
      Session Affinity:    Enabled
      Socket LB:           Enabled
      XDP Acceleration:    Disabled
      Services:
      - ClusterIP:      Enabled
      - NodePort:       Enabled (Range: 30000-32767)
      - LoadBalancer:   Enabled
      - externalIPs:    Enabled
      - HostPort:       Enabled

The new field helps to distinguish whether the datapath relies
on bpf_lxc or bpf_sock for E-W LB.

Also, I'm using socket-lb instead of host-reachable svc, as in v1.10
we are planning to rename the feature (it's used not only for reaching
services from a host nents).

Reported-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Martynas Pumputis <m@lambda.lt>